hw3/requirements.txt
transformers==4.35.2
Pillow==10.1.0
accelerate==0.24.1
bitsandbytes==0.39.0
omegaconf==2.3.0
einops==0.7.0
tensorboard==2.15.1
opencv-python==4.8.1.78
ftfy==6.1.3
kornia==0.7.0
pytorch-fid==0.3.0
pandas==2.1.3
cython
pycocotools
submitit
torch>=1.5.0
torchvision>=0.6.0
scipy
termcolor
addict
yapf
timm
jupyter
